AI_Agent_final/requirements.txt
grpcio
grpcio-tools
pandas
scikit-learn
streamlit
langchain
openai
